Started placing objects in scene
I have placed the pyramid, the tent, the bridge, and the firepit in the scene. I started doing the stars, but have only placed 4; I need to place another 4 still. Need to place trees around the map. I also added the whitebox for a super big tree I thought would be cool, but we hadn't talked about that yes so it's TBD.